+++ lib/airbrake-ruby/config.rb
@@ -5,13 +5,6 @@ module Airbrake
   # @api public
   # @since v1.0.0
   class Config
-    @instance = new
-
-    class << self
-      # @return [Config]
-      attr_accessor :instance
-    end
-
     # @return [Integer] the project identificator. This value *must* be set.
     # @api public
     attr_accessor :project_id
@@ -101,6 +94,16 @@ module Airbrake
     # @api public
     # @since v3.2.0
     attr_accessor :performance_stats_flush_period
+
+    class << self
+      # @param [Config] new_instance
+      attr_writer :instance
+
+      # @return [Config]
+      def instance
+        @instance ||= new
+      end
+    end
 
     # @param [Hash{Symbol=>Object}] user_config the hash to be used to build the
     #   config